Added the jaeger-query service that visualizes traces stored in Elasticsearch
The jaeger-query service uses a custom image containing a wait script wrapping the jaeger-query binary.

The wait script is a workaround for ensuring elastic search is ready before starting the query service. Docker does not support this natively.

TODO: figure out a way to download the jaeger-query binary in the Dockerfile, instead of saving the binary in the repo.